CI/CD Workflow
rename github workflow files. Because add a pipeline for master and next branch.
Add dockerfile for python and django app.
Add docker-compose to build dockerfile and start db. Currently for postgres as DB.
TODO Changes are required for dockerfile, docker-compose and CI Workflow
But all changes not working, because there is a confusing folder structure.