Setup Deployment enviroment on prod. Sys
Create two docker compose Files for reverse proxy (docker-compose) and app (deployScrmtl).
In  deployScrmtl is backend and db
docker-compose is traefik as reverse proxy, portaniner for  manually control for containers. All is with https and redirect. Scrmtl app is on root url. And traefik is with basic auth. The users are define in the users.txt (not part of repo)